Module lexpr::datum

S-expression values including source location.

Structs

Datum
Combines an S-expression value with location information.

ListIter
An iterator yielding the car field of a chain of cons cells.

Ref
A reference to a value and corresponding location information.

Span
The start and end for a span of text.

VectorIter
An iterator over the elements

Functions

from_reader
Parse a datum from an IO stream of S-expressions, using the default parser options.

from_reader_custom
Parse a datum from an IO stream containing a single S-expression.

from_reader_elisp
Parse a datum from an IO stream of S-expressions, using the parser options suitable for parsing Emacs Lisp.

from_slice
Parse a datum from bytes representing a single S-expressions, using the default parser options.

from_slice_custom
Parse a datum from bytes representing a single S-expression.

from_slice_elisp
Parse a datum from bytes representing a single S-expressions, using parser options suitable for Emacs Lisp.

from_str
Parse a datum from a string slice representing a single S-expressions, using the default parser options.

from_str_custom
Parse a datum from a string slice representing a single S-expression.

from_str_elisp
Parse a datum from a string slice representing a single S-expression, using parser options suitable for Emacs Lisp.