Struct lexpr::datum::Ref

pub struct Ref<'a> { /* fields omitted */ }

A reference to a value and corresponding location information.

A Ref is the generalized version of &Datum; it can not only refer a top-level, owned Datum value, but also to values recursively contained therein.

Methods

impl<'a> Ref<'a>

pub fn span(&self) -> Span

Returns the span of the referenced value.

pub fn value(&self) -> &'a Value

Returns a reference to the contained value.

pub fn list_iter(&self) -> Option<ListIter<'a>>

If the value referenced is not either a cons cell or Null, None is returned.

Note that the returned iterator has special behavior for improper lists, yielding the element after the dot after returning None the first time; see Datum::list_iter for an example.

pub fn vector_iter(&self) -> Option<VectorIter<'a>>

Returns an iterator over the elements of a vector.

If the value referenced is not a vector, None is returned.

pub fn as_pair(&self) -> Option<(Ref<'a>, Ref<'a>)>

Returns a pair of references to the fields of a cons cell.

If the value referenced is not a cons cell, None is returned.