Struct lexpr::print::Options

pub struct Options { /* fields omitted */ }

Options for printing S-expressions.

Methods

impl Options

pub fn elisp() -> Self

Construct parser options suitable for printing Emacs Lisp.

Keywords will use prefix notation (i.e. :some-keyword), the special nil value will be represented as a symbol, and booleans will be represented by nil and t.

pub fn with_keyword_syntax(self, syntax: KeywordSyntax) -> Self

Set the syntax to use for printing keywords.

pub fn with_nil_syntax(self, syntax: NilSyntax) -> Self

Set the syntax to use to print the special nil value.

pub fn with_bool_syntax(self, syntax: BoolSyntax) -> Self

Set the syntax to use to print boolean values.

pub fn with_vector_syntax(self, syntax: VectorSyntax) -> Self

Set the syntax for printing vectors.

pub fn with_bytes_syntax(self, syntax: BytesSyntax) -> Self

Set the syntax to use for printing byte vectors.

pub fn with_string_syntax(self, syntax: StringSyntax) -> Self

Set the syntax used for printing strings.

pub fn with_char_syntax(self, syntax: CharSyntax) -> Self

Set the syntax used for printing characters.